Add files via upload
Here is a video I made using Mathematica (yeah, I know). I took a noisy audio file of a normal heartbeat, with some pops, and took a spectrogram. The audio is overlayed, so we can see which features in the spectrogram correspond to heart noises and which correspond to pops, etc. This could help us determine whether spectrograms could be a useful tool for pre-processing. For instance, filtering could possibly be done differently in each time bin, where we could eliminate pops entirely if we can differentiate them from the heart sounds.